shop_metrics/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
Flask-Script
Flask-Cors
requests
mysqlclient
kafka-python